feat: Add room system and obstacle placement
In this commit, a system for changing rooms was added, where every side of the rectangle is identified by a number, top is 1, left is 2 and so on. This system made it much easier to change rooms and identify them. There are still some concerns about if this will work with bigger room systems where the id is bigger than 2. Also added an object placement system that reads from JSON files which contain the level data.